refactor a shared baseclass out of InsertRdfSourceTask
